requirements.txt
python-telegram-bot==22.5
google-auth==2.43.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.1
google-api-python-client==2.187.0
openpyxl==3.1.5
APScheduler==3.11.1
mysql-connector-python==9.5.0
requests==2.32.5
